Fix hangs on quiet interface with epoll for pcap handler + unix socket
Previously if no packets were received on an interface bpfcountd would
hang: It would neither respond to a Ctrl+c (SIGINT) nor to a unix socket
client because it would keep blocking on pcap_dispatch().

Avoid this by introducing epoll for the pcap handler and unix socket.
epoll_wait() will return when there is a new (batch of) packets from the
pcap handler, if there is a connection request from the unix socket or
if there is an interrupt from a signal handler.

This should also make the unix socket a lot more responsive.

epoll infrastructure will also help us to move the pcap filtering from
userspace to kernelspace later, to improve performance. That is epoll
allows us to easily have and handle multiple, per filter rule pcap
handlers later.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>