Introduce kernel prefilter to reduce overhead on other traffic (#10)
So far all packets are copied to userspace before any filters are
applied and before counting is performed. This can have a significant
impact on performance for unrelated traffic.

This introduces a new command line option "-F <prefilter-expr>" which
allows specifying a BPF filter expression which prefilters any packets
before being copied to userspace for further examination in bpfcountd.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>